backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beanie==1.25.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.110.0
h11==0.14.0
idna==3.10
lazy-model==0.2.0
motor==3.3.2
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.36.3
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.29.0
passlib==1.7.4
bcrypt
python-jose==3.4.0
python-multipart==0.0.20